New frontend

+++ frontend/src/pages/JournalPage.jsx
@@ -2,8 +2,15 @@ import { useState, useEffect } from "react";
 import { Form, Button, Container, Row, Col } from "react-bootstrap";
 import { api } from "../utilities/ApiUtilities";
 import { JournalEntry } from "../components/JournalEntry";
-import { useOutletContext } from "react-router";
-
+import { useOutletContext } from "react-router-dom";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faPenToSquare,
+  faTrash,
+  faRotateLeft,
+  faSquarePlus,
+  faPlus,
+} from "@fortawesome/free-solid-svg-icons";
 function JournalPage() {
   const [isViewingEntry, setIsViewingEntry] = useState(false);
   const [entrySelected, setEntrySelected] = useState(null);
@@ -42,7 +49,6 @@ function JournalPage() {
   const handleSubmit = async (event) => {
     event.preventDefault();
 
-    // Call API function here with the title and text data
     try {
       const response = await api.post("user/journal/", {
         title: title,
@@ -52,10 +58,10 @@ function JournalPage() {
       if (response.status === 201) {
         setJournalData([...journalData, response.data]);
         console.log(journalData);
-        setEntrySelected(response.data)
-        setIsViewingEntry(true)
+        setEntrySelected(response.data);
+        setIsViewingEntry(true);
 
-        // reset title and text 
+        // reset title and text
         // for reuse.
         setTitle("");
         setText("");
@@ -70,20 +76,23 @@ function JournalPage() {
   };
 
   const handleDeleteJournal = async () => {
-    try {
-      // delete from database
-      const response = await api.delete(`user/journal/${entrySelected.id}`);
-      // if successfull deletion from backend
-      if (response.status === 204) {
-        // delete from frontend useState obj
-        setJournalData((journalData) =>
-          journalData.filter((entry) => entry.id !== entrySelected.id)
-        );
-        setEntrySelected(null);
-        setIsViewingEntry(false);
+    const isConfirmed = window.confirm("Delete journal entry forever?");
+    if (isConfirmed) {
+      try {
+        // delete from database
+        const response = await api.delete(`user/journal/${entrySelected.id}`);
+        // if successfull deletion from backend
+        if (response.status === 204) {
+          // delete from frontend useState obj
+          setJournalData((journalData) =>
+            journalData.filter((entry) => entry.id !== entrySelected.id)
+          );
+          setEntrySelected(null);
+          setIsViewingEntry(false);
+        }
+      } catch (error) {
+        console.error("Error deleting data:", error);
       }
-    } catch (error) {
-      console.error("Error deleting data:", error);
     }
   };
   const putJournalEntry = async () => {
@@ -96,23 +105,23 @@ function JournalPage() {
       });
       if (response.status === 200) {
         if (response.status === 200) {
-        console.log(response.data);
-        // filter and delete journal entry with response id
-        console.log("before edit", journalData);
-          
-        const newJournalData = journalData.map((entry) => 
-          entry.id == response.data.id ? response.data : entry
-        );
-        
-        setJournalData(newJournalData);
-        setIsEditMode(false)
-        setEntrySelected(response.data)
-        setIsViewingEntry(true)
+          console.log(response.data);
+          // filter and delete journal entry with response id
+          console.log("before edit", journalData);
 
-        setText("")
-        setTitle("")
-        console.log("edit:",isEditMode, "viewing",isViewingEntry)
-      }
+          const newJournalData = journalData.map((entry) =>
+            entry.id == response.data.id ? response.data : entry
+          );
+
+          setJournalData(newJournalData);
+          setIsEditMode(false);
+          setEntrySelected(response.data);
+          setIsViewingEntry(true);
+
+          setText("");
+          setTitle("");
+          console.log("edit:", isEditMode, "viewing", isViewingEntry);
+        }
       }
     } catch (error) {
       console.error("Error editing data:", error);
@@ -127,8 +136,8 @@ function JournalPage() {
     console.log(`editmode:${isEditMode} viewmode:${isViewingEntry}`);
   };
   const handleCreateButton = () => {
-    setIsViewingEntry(false)
-    setIsEditMode(false)
+    setIsViewingEntry(false);
+    setIsEditMode(false);
   };
   const undoEditMode = () => {
     setIsViewingEntry(true);
@@ -137,74 +146,111 @@ function JournalPage() {
 
   return (
     <>
-      <h2>Journal</h2>
-      <div className={`container fade-in-out ${isVisible ? "" : "fade-out"}`}>
+      <div className={`fade-in-out ${isVisible ? "" : "fade-out"}`}>
         <div className="row">
           <div
-            className={`col-2 scrollable fade-in-out ${
-              isVisible ? "" : "fade-out"
-            }`}
-            style={{ backgroundColor: "gray" }}
+            style={{ border: "none" }}
+            className={`col-12 col-md-2 fade-in-out ${isVisible ? "" : "fade-out"}`}
           >
-            {
-            journalData !== null &&
-            journalData !== undefined &&
-            journalData.length > 0
-            ? (
-              journalData.map((journal) => (
-                <JournalEntry
-                  key={journal.id}
-                  id={journal.id}
-                  title={journal.title}
-                  text={journal.text}
-                  isViewingEntry={isViewingEntry}
-                  entrySelected={entrySelected}
-                  setIsViewingEntry={setIsViewingEntry}
-                  setEntrySelected={setEntrySelected}
-                />
-              ))
-            ) : (
-              <h1>empty</h1>
-            )}
+            <Container style={{}}>
+              <div style={{textAlign:"center"}}>
+              <h1 style={{display:"inline"}}>Entries</h1><FontAwesomeIcon icon={faPlus} className="plus-icon" onClick={handleCreateButton}/>
+              </div>
+              <div style={{textAlign:"center"}}>
+              
+              </div>
+              <div
+              className="scrollable-box"
+                style={{
+                  // overflowY: "auto",
+                  width: "11vw",
+                  position: "absolute",
+                  height: "250px",
+            
+                    wordBreak:"break-all"
+                }}
+              >
+                <Col
+                
+                  style={{
+                    position: "relative",
+                    height: "70vh",
+                    paddingLeft: "15px",
+                    display: "flex",
+                    flexDirection: "column",
+                    alignItems: "center",
+                    wordBreak:"break-all"
+                  }}
+                >
+                  {/* <Row style={{ fontSize: "34px" }}>Journal</Row> */}
+                  {journalData.length > 0 ? (
+                    journalData.map((journal) => (
+                      <JournalEntry
+                        key={journal.id}
+                        id={journal.id}
+                        title={journal.title}
+                        text={journal.text}
+                        isViewingEntry={isViewingEntry}
+                        entrySelected={entrySelected}
+                        setIsViewingEntry={setIsViewingEntry}
+                        setEntrySelected={setEntrySelected}
+                      />
+                    ))
+                  ) : (
+                    <>
+                      <h1>empty</h1>
+                    </>
+                  )}
+                </Col>
+              </div>
+            </Container>
           </div>
-          <div className="col-7">
+
+          <div className="col-12 col-md-7">
             <Container>
               <Row className="justify-content-md-center mt-5">
                 <Col xs={12} md={6}>
                   {isViewingEntry || isEditMode ? (
                     isViewingEntry ? (
                       <>
-                        <div>
+                        <div className="entry">
+                          <Row>
+                            <Col
+                              md={{ span: 6, offset: 0 }}
+                              className="text-left"
+                            ></Col>
+                            <Col
+                              md={{ span: 6, offset: 0 }}
+                              className="text-right"
+                            ></Col>
+                          </Row>
+                          <div style={{wordWrap:"break-word"}}>
                           <h1>{entrySelected.title}</h1>
                           <p>{entrySelected.text}</p>
-                        </div>
-                        <div>
-                          <Button
-                            variant="secondary"
-                            onClick={handleEditButton}
-                          >
-                            edit
-                          </Button>
-                          <Button
-                            variant="danger"
-                            onClick={handleDeleteJournal}
-                          >
-                            delete
-                          </Button>
-                          <Button
-                            variant="success"
-                            onClick={handleCreateButton}
-                          >
-                            create new
-                          </Button>
+                          </div>
+                          <Row>
+                            
+                            <Col md={6}>
+                              <FontAwesomeIcon
+                                icon={faTrash}
+                                onClick={handleDeleteJournal}
+                                className="journal-icon delete-icon"
+                              />
+                              <FontAwesomeIcon
+                                icon={faPenToSquare}
+                                onClick={handleEditButton}
+                                className="journal-icon edit-icon"
+                              />
+                            </Col>
+                          </Row>
                         </div>
                       </>
                     ) : (
                       // render editmode
                       <>
-                        <Form onSubmit={handleSubmit}>
+                       
+                        <Form onSubmit={handleSubmit} className="journal-form">
                           <Form.Group controlId="title">
-                            <Form.Label>Title:</Form.Label>
                             <Form.Control
                               type="text"
                               value={title}
@@ -213,7 +259,7 @@ function JournalPage() {
                           </Form.Group>
 
                           <Form.Group controlId="text">
-                            <Form.Label>Text:</Form.Label>
+                            <Form.Label></Form.Label>
                             <Form.Control
                               as="textarea"
                               rows={4}
@@ -221,42 +267,54 @@ function JournalPage() {
                               onChange={handleTextChange}
                             />
                           </Form.Group>
-                          <Button
-                            variant="primary"
-                            type="submit"
-                            onClick={putJournalEntry}
-                          >
-                            Submit Edit
-                          </Button>
-                          <Button varient="secondary" onClick={undoEditMode}>
-                            undo
-                          </Button>
+
+                          <div className="d-flex justify-content-between">
+                            <Button
+                              variant="outline-light"
+                              type="submit"
+                              onClick={putJournalEntry}
+                            >
+                              save
+                            </Button>
+
+                            <FontAwesomeIcon
+                              icon={faRotateLeft}
+                              onClick={undoEditMode}
+                            />
+                          </div>
                         </Form>
                       </>
                     )
                   ) : (
-                    <Form onSubmit={handleSubmit}>
+                    <Form
+                      onSubmit={handleSubmit}
+                      className="journal-form"
+                      autoComplete="off"
+                    >
                       <Form.Group controlId="title">
-                        <Form.Label>Title:</Form.Label>
                         <Form.Control
                           type="text"
                           value={title}
                           onChange={handleTitleChange}
+                          placeholder="Journal Title"
                         />
                       </Form.Group>
 
                       <Form.Group controlId="text">
-                        <Form.Label>Text:</Form.Label>
+                        <Form.Label></Form.Label>
                         <Form.Control
                           as="textarea"
                           rows={4}
                           value={text}
                           onChange={handleTextChange}
+                          placeholder="Write your journal entry"
                         />
                       </Form.Group>
-                      <Button variant="primary" type="submit">
+                      <div style={{textAlign:"center"}}>
+                      <button className="journal-button"  type="submit">
                         Submit
-                      </Button>
+                      </button>
+                      </div>
                     </Form>
                   )}
                 </Col>
